+++ docs/.docs_source_mirror_stamp.json
@@ -2,8 +2,8 @@
   "file_count": 166,
   "format_version": 1,
   "managed_target": "docs/source",
-  "source_digest_sha256": "eaf2dd49a0cc28ae9545eec2683c05394795fba59dd3c5a03add20bff223e62d",
+  "source_digest_sha256": "732c2d1bba01a65c9e370a1f9ab9747b416fcce0150971a636039e2e49ca5d22",
   "source_hint": "../thales_agilab/docs/source",
   "sync_tool": "tools/sync_docs_source.py",
-  "target_digest_sha256": "eaf2dd49a0cc28ae9545eec2683c05394795fba59dd3c5a03add20bff223e62d"
+  "target_digest_sha256": "732c2d1bba01a65c9e370a1f9ab9747b416fcce0150971a636039e2e49ca5d22"
 }
