+++ src/agilab/about_page/__init__.py
@@ -1,0 +1,2 @@
+"""Implementation modules for the AGILab About page."""
+
