+++ src/agilab/lib/agi-gui/test/test_agi_gui_package.py
@@ -1,8 +1,10 @@
 from __future__ import annotations
 
 import importlib
+import sys
 import tomllib
 from pathlib import Path
+from types import ModuleType
 
 
 AGI_GUI_ROOT = Path(__file__).resolve().parents[1]
@@ -30,3 +32,37 @@ def test_agi_gui_exports_file_picker_helpers() -> None:
     assert callable(module.confirm_button)
     assert callable(module.status_container)
     assert callable(module.toast)
+
+
+def test_compat_alias_exposes_source_module(monkeypatch) -> None:
+    compat = importlib.import_module("agi_gui._compat")
+    source = ModuleType("demo_source_module")
+    monkeypatch.setitem(sys.modules, "demo_source_module", source)
+
+    result = compat.alias_agi_env_module("agi_gui.demo_alias", "demo_source_module")
+
+    assert result is source
+    assert sys.modules["agi_gui.demo_alias"] is source
+
+
+def test_compatibility_proxy_modules_import_from_agi_env() -> None:
+    proxy_names = [
+        "pagelib",
+        "pagelib_data_support",
+        "pagelib_execution_support",
+        "pagelib_navigation_support",
+        "pagelib_preview_support",
+        "pagelib_project_support",
+        "pagelib_resource_support",
+        "pagelib_runtime_support",
+        "pagelib_selection_support",
+        "pagelib_session_support",
+        "streamlit_args",
+        "ui_docs_support",
+        "ui_state_support",
+        "ui_support",
+    ]
+
+    for proxy_name in proxy_names:
+        module = importlib.import_module(f"agi_gui.{proxy_name}")
+        assert module.__name__ == f"agi_env.{proxy_name}"
