+++ src/agilab/resources/config.toml
@@ -1,6 +1,6 @@
 [theme]
 base = "light"
-primaryColor = "#0D47A1"
+primaryColor = "#4A90E2"
 backgroundColor = "#F2F6FF"
 secondaryBackgroundColor = "#FFFFFF"
 textColor = "#102A43"
